apps/ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
langchain-huggingface
sentence-transformers
pypdf
pandas
pydantic
langchain-groq
python-dotenv
duckduckgo-search
